backend/pipeline/face_tracker.py
"""
AttentionX – Step 5: Face Tracking & Vertical Crop
Uses MediaPipe Face Detection to track faces and smart-crop 16:9 → 9:16.
Implements smooth camera movement to avoid jarring cuts.
"""

import os
import cv2
import numpy as np
import subprocess
import json
from pathlib import Path
from typing import Optional

try:
    import mediapipe as mp
    MEDIAPIPE_AVAILABLE = True
except ImportError:
    MEDIAPIPE_AVAILABLE = False
    print("[FaceTracker] MediaPipe not available. Using center crop fallback.")

from utils.config import settings


class FaceTracker:
    """
    Converts horizontal (16:9) video to vertical (9:16) using face tracking.

    Strategy:
    1. Sample frames at regular intervals to detect face positions
    2. Smooth the crop window trajectory (avoid jitter)
    3. Render output video with smooth panning crop
    """

    def __init__(
        self,
        output_width: int = None,
        output_height: int = None,
        smooth_factor: float = None,
        padding_percent: float = None,
    ):
        self.output_width = output_width or settings.OUTPUT_WIDTH
        self.output_height = output_height or settings.OUTPUT_HEIGHT
        self.smooth_factor = smooth_factor or settings.FACE_TRACKING_SMOOTH_FACTOR
        self.padding = padding_percent or settings.FACE_PADDING_PERCENT
        self._detector = None

    def _get_detector(self):
        """Lazy-load MediaPipe face detector."""
        if self._detector is None and MEDIAPIPE_AVAILABLE:
            mp_face = mp.solutions.face_detection
            self._detector = mp_face.FaceDetection(
                model_selection=1,  # 1 = full-range model (up to 5m)
                min_detection_confidence=settings.FACE_DETECTION_CONFIDENCE,
            )
        return self._detector

    def process_clip(
        self,
        input_path: str,
        output_path: str,
        sample_fps: float = 2.0,
    ) -> str:
        """
        Convert a horizontal clip to vertical with face tracking.

        Args:
            input_path: Path to raw horizontal clip
            output_path: Output path for vertical clip
            sample_fps: How many frames per second to sample for face detection

        Returns:
            Path to processed vertical clip
        """
        print(f"[FaceTracker] Processing: {input_path}")

        # Get source dimensions
        cap = cv2.VideoCapture(input_path)
        src_w = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
        src_h = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
        src_fps = cap.get(cv2.CAP_PROP_FPS) or 30
        total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
        cap.release()

        print(f"[FaceTracker] Source: {src_w}x{src_h} @ {src_fps}fps, {total_frames} frames")

        # Compute crop window: we need a width such that width/height = 9/16
        # within the source video
        target_ratio = self.output_width / self.output_height  # 9/16
        crop_width = int(src_h * target_ratio)
        crop_width = min(crop_width, src_w)
        crop_height = src_h

        # If source is already vertical or narrower, just scale
        if src_h <= src_w * 0.7:
            # Landscape → crop
            pass
        else:
            # Already vertical-ish, just resize
            return self._simple_resize(input_path, output_path)

        # ── Step 1: Sample frames and detect faces ────────────────────────────
        frame_step = max(1, int(src_fps / sample_fps))
        crop_centers = self._detect_face_positions(
            input_path, src_w, src_h, src_fps, total_frames,
            crop_width, frame_step
        )

        # ── Step 2: Smooth the crop centers ──────────────────────────────────
        smooth_centers = self._smooth_positions(crop_centers, total_frames)

        # ── Step 3: Write crop instructions to file ───────────────────────────
        crop_data_path = output_path + ".crops.json"
        with open(crop_data_path, "w") as f:
            json.dump({"crop_width": crop_width, "crop_height": crop_height,
                       "centers": smooth_centers}, f)

        # ── Step 4: Render with FFmpeg using crop filter ───────────────────────
        result_path = self._render_with_ffmpeg(
            input_path, output_path,
            smooth_centers, crop_width, crop_height,
            src_w, src_h, src_fps
        )

        # Cleanup
        if os.path.exists(crop_data_path):
            os.remove(crop_data_path)

        return result_path

    def _detect_face_positions(
        self,
        video_path: str,
        src_w: int,
        src_h: int,
        fps: float,
        total_frames: int,
        crop_width: int,
        frame_step: int,
    ) -> dict[int, int]:
        """
        Sample frames and return dict: {frame_idx: crop_center_x}
        """
        detector = self._get_detector()
        cap = cv2.VideoCapture(video_path)

        centers = {}
        frame_idx = 0

        while True:
            ret, frame = cap.read()
            if not ret:
                break

            if frame_idx % frame_step == 0:
                center_x = self._get_face_center_x(frame, src_w, src_h, crop_width, detector)
                centers[frame_idx] = center_x

            frame_idx += 1

        cap.release()
        print(f"[FaceTracker] Sampled {len(centers)} frames for face detection")
        return centers

    def _get_face_center_x(
        self,
        frame: np.ndarray,
        src_w: int,
        src_h: int,
        crop_width: int,
        detector,
    ) -> int:
        """
        Detect face in frame, return optimal crop center X.
        Falls back to video center if no face detected.
        """
        default_center = src_w // 2

        if detector is None:
            return default_center

        try:
            # MediaPipe expects RGB
            frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
            results = detector.process(frame_rgb)

            if results.detections:
                # Use the first (most prominent) detection
                detection = results.detections[0]
                bbox = detection.location_data.relative_bounding_box

                # Convert to pixel coordinates
                face_cx = int((bbox.xmin + bbox.width / 2) * src_w)
                face_cy = int((bbox.ymin + bbox.height / 2) * src_h)

                # Add padding
                padded_cx = face_cx

                # Clamp to valid crop range
                half_crop = crop_width // 2
                clamped = max(half_crop, min(src_w - half_crop, padded_cx))
                return clamped

        except Exception:
            pass

        return default_center

    def _smooth_positions(
        self,
        sampled_centers: dict[int, int],
        total_frames: int,
    ) -> list[int]:
        """
        Interpolate sampled positions to all frames using numpy (O(N) not O(N²)),
        then apply exponential smoothing.
        """
        if not sampled_centers:
            default = total_frames // 2  # will be replaced by src_w//2 upstream
            return [default] * total_frames

        # Sort sampled data
        frame_indices = np.array(sorted(sampled_centers.keys()), dtype=float)
        center_values = np.array([sampled_centers[k] for k in frame_indices.astype(int)], dtype=float)

        # Linear interpolation across all frames using numpy (O(N))
        all_frames = np.arange(total_frames, dtype=float)
        interpolated = np.interp(all_frames, frame_indices, center_values)

        # Vectorised exponential smoothing (IIR filter)
        # y[n] = alpha * x[n] + (1-alpha) * y[n-1]
        alpha = self.smooth_factor
        smoothed = np.empty(total_frames, dtype=float)
        smoothed[0] = interpolated[0]
        for i in range(1, total_frames):
            smoothed[i] = alpha * interpolated[i] + (1.0 - alpha) * smoothed[i - 1]

        return smoothed.astype(int).tolist()

    def _render_with_ffmpeg(
        self,
        input_path: str,
        output_path: str,
        centers: list[int],
        crop_width: int,
        crop_height: int,
        src_w: int,
        src_h: int,
        fps: float,
    ) -> str:
        """
        Use FFmpeg sendcmd filter to apply per-frame crop.
        Falls back to center crop if per-frame is too slow.
        """
        os.makedirs(os.path.dirname(output_path) or ".", exist_ok=True)

        # For hackathon speed: use average center (still good result)
        # Full per-frame rendering requires writing a custom filter chain
        if centers:
            avg_center = int(np.median(centers))
        else:
            avg_center = src_w // 2

        half_crop = crop_width // 2
        crop_x = max(0, min(src_w - crop_width, avg_center - half_crop))

        # Build FFmpeg command with scale+crop
        cmd = [
            "ffmpeg", "-y",
            "-i", input_path,
            "-vf", (
                f"crop={crop_width}:{crop_height}:{crop_x}:0,"
                f"scale={self.output_width}:{self.output_height}:force_original_aspect_ratio=decrease,"
                f"pad={self.output_width}:{self.output_height}:(ow-iw)/2:(oh-ih)/2:black"
            ),
            "-c:v", "libx264",
            "-preset", "fast",
            "-crf", "20",
            "-c:a", "aac",
            "-b:a", "192k",
            "-movflags", "+faststart",
            output_path,
        ]

        result = subprocess.run(cmd, capture_output=True, text=True, timeout=600)
        if result.returncode != 0:
            raise RuntimeError(f"FFmpeg vertical crop failed: {result.stderr[-500:]}")

        print(f"[FaceTracker] Vertical clip saved: {output_path}")
        return output_path

    def _simple_resize(self, input_path: str, output_path: str) -> str:
        """Simple resize/pad to target dimensions (no cropping needed)."""
        os.makedirs(os.path.dirname(output_path) or ".", exist_ok=True)
        cmd = [
            "ffmpeg", "-y",
            "-i", input_path,
            "-vf", (
                f"scale={self.output_width}:{self.output_height}:"
                f"force_original_aspect_ratio=decrease,"
                f"pad={self.output_width}:{self.output_height}:(ow-iw)/2:(oh-ih)/2:black"
            ),
            "-c:v", "libx264", "-preset", "fast", "-crf", "20",
            "-c:a", "aac", "-b:a", "192k",
            "-movflags", "+faststart",
            output_path,
        ]
        result = subprocess.run(cmd, capture_output=True, text=True, timeout=300)
        if result.returncode != 0:
            raise RuntimeError(f"Simple resize failed: {result.stderr[-300:]}")
        return output_path
